# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games-API
    version: 1.0.0
    description: >-
        ### This is a Swagger UI for our Games API  


        ### Error handling

        Errors are presented using the HTTP status codes 400-599. The response include a more detailed
        explanation about the error in the body.


        ```

        {
          "error": "Cleartext explanation of the error that occurred"
        }

        ```


        ### Generic API errors

        These error responses can be returned by any of the described endpoints.


        | HTTP status  |  Description |

        |---|---|

        | 400| Bad input parameters or invalid request. More information can be found in the response body.

        | 401| The consumer does not have access to the requested resource or the provided token is not valid.

        | 404| The requested resource can not be found or the user does not have access to the resource.

        | 5XX| A server error has occurred while processing the request.
paths:
    /games:
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of games:
                                    value:
                                        - Witcher 3
                                        - Cyberpunk 2077
                                        - Minecraft
                    description: 200 OK
            summary: Get a list of all games
            description: Return a list of all games in API
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Full details of a given game:
                                    value: "    {id: 1, name: \"Witcher 3\", price: 29.99},\r\n    {id: 2, name: \"Cyberpunk 2077\", price: 59.99},\r\n    {id: 3, name: \"Minecraft\", price: 26.99},\r\n    {id: 4, name: \"Counter-Strike: Global Offensice\", price: 0},\r\n    {id: 5, name: \"Roblox\", price: 0},\r\n    {id: 6, name: \"Grand Theft Auto V\", price: 29.99},\r\n    {id: 7, name: \"Valorant\", price: 0},\r\n    {id: 8, name: \"Forza Horizon 5\", price: 59.99}"
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            examples:
                                Game /w+ was not found:
                                    value: ' Game \\w+ was not found in game array'
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of a games
            description: Returns full details of a given id game
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.